fix(platform): 修复跨平台兼容性问题
refactor(memory): 重构内存模块头文件引用路径

feat(strategy): 添加LLM服务前向声明

fix(vector_store): 修复SQLite绑定参数类型转换问题

feat(retrieval): 添加吸引子网络和检索缓存前向声明

refactor(execution): 清理执行单元头文件引用路径

feat(heapstore): 添加会话查询接口和JSON导出功能

fix(ipc): 修复Windows平台IPC兼容性和消息大小检查

refactor(monitoring): 移除observability依赖并添加cJSON条件编译

refactor(memory_optimizer): 移除observability依赖并添加cJSON条件编译

chore(vcpkg): 简化Windows依赖配置

docs(issue_template): 完善Bug报告模板和标签描述

refactor(token_counter): 使用平台互斥锁替代自定义实现

refactor(gateway_d): 重构CMake构建配置和依赖管理

docs(bug_report): 完善Bug报告模板并添加架构影响分析

+++ atoms/memoryrovol/src/vector_store.c
@@ -165,7 +165,7 @@ agentos_error_t agentos_vector_store_put(
 
     /* 绑定参数 */
     sqlite3_bind_text(stmt, 1, record_id, -1, SQLITE_STATIC);
-    sqlite3_bind_blob(stmt, 2, vector, dim * sizeof(float), SQLITE_STATIC);
+    sqlite3_bind_blob(stmt, 2, vector, (int)(dim * sizeof(float)), SQLITE_STATIC);
     sqlite3_bind_int(stmt, 3, (int)dim);
 
     /* 执行 */
